refactor downloader
